Optimizer (#8)
This will effectively attempt to execute each transformation during the compilation stage.

If the expression returns a SourceMissingError, it will recursively optimize each input in the same way. If it returns a value, it will be replaced by that value, if it returns any other error, that error is passed along as a compiler error.

This helps do a few trivial optimizations, which may be helpful when running more complex expressions. As this is all done during compile time, the cost doesn't really matter, so this should be an objective benefit.