Simplify kuiper
This removes the entire "Program" part of the language, which is no
longer necessary since we have lambda expressions. Neither the CLI nor
the python wrapper uses it.

This lets us remove the "ID" concept as well, which doesn't really make
sense without the program blocks.

The tests relied heavily on this old behavior, so many are removed, and
others are modified heavily.

The diff is pleasant at least.