Remove playground
It was causing issues for renovate, and if we wanted to include it in
the build we would have to compile the SDK and a bunch of other stuff on
build, which wasn't very necessary for something that isn't a great
example anyway. We can replace it with something that only dependes on
kuiper later.