Make `try_float` understand commas (#85)
I've been looking at a few different REST APIs when doing the design doc
for REST in Pluto, and a couple of them formats their floats with commas
instead of dots for decimals. F.ex. NordPool will format their responses
as

``` json
{
	...
	"Value": "815,43",
	...
}
```

which is pretty dumb, but something we should handle.