Improve error spans in a few cases
For an operator, the span shouldn't cover both right and left sides of
the statement, because this is rarely correct.

Similar for selectors and postfix functions. In general, covering a
non-enclosed Term is dangerous because it might be the entire rest of
the expression, so we avoid that to avoid producing useless spans.