Automatically generate list of builtins
The list of builtins were defined several places:

 * in kuiper_lang (source of truth)
 * in kuiper_cli (for auto completion)
 * in kuiper_lezer (for auto completion and help windows)
 * in documentation

Now, it's defined in a yaml file, and cli, lezer and docs are all auto
created from this file. It also cross checks with the match statement
in kuiper_lang to make sure the lists stay in-sync.

The only manual process now is to copy the `built_in_functions.md` file
and make a PR on the docs repo.