Some perf tuning (#175)
Improve the try_* methods a bit, and avoid a few string clones. Some of these the compiler already managed to optimize out, but it is nice to be consistent about it.

This also adds benchmarking functions handling rockwell and cognite formats. They cover a fairly narrow part of the language, but the parts they cover are likely quite important.