Add option to limit number of operations in a kuiper expression  (#210)
This ended up being fairly complicated. Turns out, the borrow checker can take a few shortcuts when only working with immutable references. While semantically having the state be mutable is coherent, it requires being a bit more strict with the borrow checker. I even ran into a fairly significant limitation.

Once the next generation borrow checker is implemented, a lot of this will likely be better, but that is still quite a ways off.