Substring function
Requires some finesse actually, since by default slicing into a string
in rust slices on bytes, not characters, and will panic if you split a
utf-8 codepoint.

I landed on the most error-tolerant version of substring I could think
of, where out of bounds is clamped to the bounds of the string,
effectively.

Javascript bizarrely swaps start and end if start > end, which IMO makes
no sense at all, so I didn't do that. If the range is empty, then return
nothing.