Add tail function
Add a function to get the last element(s) of a list.

Based on a hub post where someone were subscribing to `some/prefix/#`,
and wanted the last level of the topic (which was the sensor id). This
in combination with the upcoming `slice` would make that very easy to
do: `context.topic.slice("/").tail()`