Refactor result (#323)
This changes ResolveResult into our own type, instead of being an alias for Cow. The disadvantage of this is that we need to reimplement a fair few of the traits implemented on Cow, but the advantage is that we can tailor it more specifically to our needs, and also easily add new methods on it. This changes the old global get_x_from_value methods to be methods on ResolveResult directly, which leads to much more readable code, imo.