Support negative numbers in array indexing (#554)
This is handy, and something a lot of users expect. It lets you do [1, 2, 3][-1] == 3, like python and other languages. Current workarounds are a pain, and this is something we can painlessly support.

Also wrote some general tests for selectors. We have a few, and they are used in lots of other tests (being pretty core to the language), but having some dedicated tests is good.