Make concat in object expression join conservatively
This is technically a bit of a breaking change, but the current behavior
is _very_ unintuitive, and I believe it is worth it.

This makes it so that the `...` operator only adds elements if they are
not already in the object, which is how rust treats these. The current
behavior (replacing the value), is really unintuitive.

Explicitly setting the value still overwrites, so

```
...{
    "a": 1,
},
"a": 2
```

still yields `{ "a": 2 }`, which makes this change _much_ less breaking.
You'd need to be relying on the fact that you have a useless asignment
for this change to cause issues.