Add a framework for bulk testing sample expressions (#564)
I want to add a large bank of known-to-be-good expressions that we
evaluate the compiler on. The idea is to have a collection of
expressions that we compile (and optionally run) as part of the test
suite to catch possible regressions. We can collect these over time, and
also base them on what we see actual users do.

The idea is very similar to what they do for the
[Rust language itself](https://github.com/rust-lang/rust/tree/master/tests/ui/parser).

This PR mostly just adds the scaffolding to make it work, and a few test
cases just to show the concept. I'll go hunting for test cases later.