Initial commit for type checking framework (#616)
* Initial commit for type checking framework

I've previously implemented a full type checking framework for Kuiper,
but the PR is a scary 2.5k lines of code, which would be hard to review
in the best of times. This is the first of _many_ PRs to implement type
checking for the language.

First, I'll implement the `Type` enum itself, the next few PRs will
implement sequences, objects, logic for checking compatibility between
types, and logic for creating unions of types.

Second, I'll start working my way through the entire language, implement
"type space" resolvers for every expression type. This is a lot of work,
but I've already done the vast majority of it, so it will not take that
much time to write.

Third, we'll make this available when calling `compile_expression`, but
we won't enable type checking immediately. First we'll write some simple
code in pluto to load all mappings from the database and run type
checking on all of them, to check for incompatibility. This is useful
anyway, since an incompatible mapping stored in the database indicates a
regression in the language, which would be nice to catch in the future.

Finally, we'll make this a standard feature. At this point we can look
at interesting options:

 - Checking in the frontend whether mapping expressions can be valid.
 - Analyzing the output of mapping expressions to identify where they
   are writing to, thus allowing us to create static lineage. I suspect
   this will work well in a lot of cases.
 - Improving type checking further to allow using it for optimization,
   and catching more errors.

For backwards compatibility I aim to make it so that the "default" type
checking is just "Accept `Any` return `Any`".

* Clippy

* Review comment