Union creation method
This is fairly complex, and there are a lot of possible tradeofs we can
make. I've decided that for now we'll merge constant and generic types,
i.e. union(string, "hello") = string, and ignore objects, so union({
"foo": "bar" }, { "foo": String }) is just a union object. We may change
this later depending on need. I've written methods to union objects and
arrays with each other, which is handy for things like try_as_array and
try_as_object.