Framework for custom input values
This comes from a proposal to add new values to the `context` object in
pluto. I'm concerned that as the context object grows, we risk spending
a lot of time allocating large JSON blobs for each transform, even if
the actual payloads are tiny.

This is an attempt at a solution for that, without parametrizing the
entire expression over this. This _does_ come with a performance hit.
The benchmarks are probably quite source-selecting heavy, but it looks
like this may cost as much as 10% performance in the worst case, likely
due to the inability to inline.