Framework for custom input values (#626)
This comes from a proposal to add new values to the `context` object in pluto. I'm concerned that as the context object grows, we risk spending a lot of time allocating large JSON blobs for each transform, even if the actual payloads are tiny.

This is an attempt at a solution for that, without parametrizing the entire expression over this. This _does_ come with a performance hit. The benchmarks are probably quite source-selecting heavy, ~~but it looks like this may cost as much as 6-7% performance hit in the worst case, likely due to the inability to inline.~~ actually I managed to save pretty much that entire performance cost by avoiding the call to `keys` in completions. It's possible that would save a little bit without this feature as well, but that appears to have been a significant part of the added cost...

It's hard to say whether this is really worth it, but implementation wise it was fairly easy to do.