Types framework: is_assignable_to (#638)
This checks whether a type _could_ be assigned to a different value,
meaning that there exists a narrowing of `self` such that self
could be assigned to `other`.

This is a little hairy, and ends up being very relaxed, but it's pretty
much the strictest we can be for type checking. Anything more strict
than this would require type assertions in the language itself.