Basic code coverage (#648)
This is just for the main rust library. Getting proper coverage for the other modules might be hard since those are invoked indirectly through other languages. For python, for example, there isn't really any python code at all, it's all a dynamic library.

In practice, those modules are a tiny, tiny part of the total code.